requirements.txt
boto3
dlib
numpy
face_recognition
opencv-python
ffmpeg-python
